# :rocket: OpenMP support for `cmatrix()`