## :books: What?

* **OpenMP support for confusion matrix:** The `cmatrix()` now supports OpenMP. As *most* classification functions uses `cmatrix()` in the backend, the OpenMP is now supported for most classification functions.

## :eyes: Showcase

``` r
# 1) set seed
rm(list = ls()); invisible(gc());

# 2) define values
# for classes
set.seed(1903)
actual <- factor(sample(letters[1:3], 1e7, TRUE))
predicted <- factor(sample(letters[1:3], 1e7, TRUE))

# 3) benchmark with OpenMP
SLmetrics::setUseOpenMP(TRUE)
#> OpenMP usage set to: enabled

bench::mark(
  `{With OpenMP}`  = SLmetrics::cmatrix(actual, predicted)
)
#> # A tibble: 1 × 6
#>   expression         min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>    <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 {With OpenMP}    556µs    587µs     1463.    2.51KB     2.00

# 4) benchmark without OpenMP
SLmetrics::setUseOpenMP(FALSE)
#> OpenMP usage set to: disabled

bench::mark(
  `{Without OpenMP}`  = SLmetrics::cmatrix(actual, predicted)
)
#> # A tibble: 1 × 6
#>   expression            min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>       <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 {Without OpenMP}   6.02ms   6.07ms      165.        0B        0
```

<sup>Created on 2025-01-10 with [reprex v2.1.1](https://reprex.tidyverse.org)</sup>